Restructure tests to live alongside their code.
This allows for easier reviewing. The tests will be directly underneath
the code change in the GitHub UI so you can validate things are properly
tested easier.
