+++ portal/views/__init__.py
@@ -1,0 +1,11 @@
+"""
+Import submodules to allow for nicer code separation.
+"""
+# pylint: disable=wildcard-import
+from portal.views.activation import *
+from portal.views.ccxcon_api import *
+from portal.views.login import *
+from portal.views.product_api import *
+from portal.views.registration import *
+from portal.views.webhooks import *
+from portal.views.webpack import *
