+++ portal/views/__init_test.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+"""
+Testing of the module level stuff itself
+"""
+import unittest
+
+import semantic_version
+
+
+class TestModule(unittest.TestCase):
+    """
+    Test core module features, like asserting the version
+    and making sure we are exposing our classes.
+    """
+
+    @staticmethod
+    def test_version():
+        """
+        Verify we have a valid semantic version
+        """
+        import portal
+        semantic_version.Version(portal.__version__)
