+++ portal/views/base.py
