+++ static/js/global_init.js
@@ -1,5 +1,5 @@
 // Define globals we would usually get from Django
 global.SETTINGS = {
   ccxconApi: 'ccxcon/',
-  isAuthenticated: false,
+  isAuthenticated: false
 };
