+++ static/js/mocha.opts
