+++ tox.ini
@@ -22,7 +22,7 @@ whitelist_externals =
 commands =
     npm install --no-bin-links --prefix {toxinidir}
     node {toxinidir}/node_modules/eslint/bin/eslint.js {toxinidir}/static/js
-    node {toxinidir}/node_modules/mocha/bin/mocha
+    node {toxinidir}/node_modules/mocha/bin/mocha --opts static/js/mocha.opts static/js/global_init.js "static/**/*_test.js"
 
 whitelist_externals =
     npm
